Current version converges correctly
It seems that a double counting in the cell lists was giving an error in
the full simulation, accumulating errors throughout. Now the problem
seems to be solved, as the simulation can now reach a target maximum
strain. The simple version of the code is a brute-force approach to
check whether the code was correct or not.